+++ crates/xtask-bump-check/src/main.rs
@@ -8,7 +8,7 @@ fn main() {
     let cli = xtask::cli();
     let matches = cli.get_matches();
 
-    let mut gctx = cargo::util::context::GlobalContext::default().unwrap_or_else(|e| {
+    let mut gctx = cargo::context::GlobalContext::default().unwrap_or_else(|e| {
         let mut eval = cargo_util_terminal::Shell::new();
         cargo::exit_with_error(e.into(), &mut eval)
     });
