+++ src/compiler/build_context/mod.rs
@@ -5,12 +5,12 @@ use crate::compiler::CompileKind;
 use crate::compiler::Unit;
 use crate::compiler::UnitIndex;
 use crate::compiler::unit_graph::UnitGraph;
+use crate::context::GlobalContext;
 use crate::core::PackageSet;
 use crate::core::Workspace;
 use crate::core::dependency::DepKind;
 use crate::core::profiles::Profiles;
 use crate::util::Rustc;
-use crate::util::context::GlobalContext;
 use crate::util::data_structures::{HashMap, HashSet};
 use crate::util::errors::CargoResult;
 use crate::util::interning::InternedString;
