+++ src/compiler/job_queue/mod.rs
@@ -140,12 +140,12 @@ use super::timings::SectionTiming;
 use super::timings::Timings;
 use crate::compiler::descriptive_pkg_name;
 use crate::compiler::future_incompat::{self, FutureBreakageItem, FutureIncompatReportPackage};
+use crate::context::WarningHandling;
 use crate::core::{PackageId, TargetKind};
 use crate::diagnostics::GlobalDiagnosticStats;
 use crate::diagnostics::rules::unused_dependencies;
 use crate::resolver::ResolveBehavior;
 use crate::util::CargoResult;
-use crate::util::context::WarningHandling;
 use crate::util::diagnostic_server::{self, DiagnosticPrinter};
 use crate::util::errors::AlreadyPrintedError;
 use crate::util::interning::InternedString;
