+++ src/core/profiles.rs
@@ -22,6 +22,7 @@
 //! The algorithm happens within [`ProfileMaker::get_profile`].
 
 use crate::compiler::{CompileKind, CompileTarget, Unit};
+use crate::context;
 use crate::core::Feature;
 use crate::core::dependency::Artifact;
 use crate::core::{PackageId, PackageIdSpec, PackageIdSpecQuery, Target, Workspace};
@@ -30,7 +31,7 @@ use crate::resolver::features::FeaturesFor;
 use crate::util::data_structures::{HashMap, HashSet};
 use crate::util::interning::InternedString;
 use crate::util::toml::validate_profile;
-use crate::util::{CargoResult, GlobalContext, closest_msg, context};
+use crate::util::{CargoResult, GlobalContext, closest_msg};
 use anyhow::{Context as _, bail};
 use cargo_util::is_ci;
 use cargo_util_schemas::manifest::TomlTrimPaths;
