+++ src/diagnostics/lint.rs
@@ -4,9 +4,9 @@ use std::fmt::Display;
 use cargo_util_schemas::manifest;
 use cargo_util_terminal::report::Level;
 
+use crate::context::WarningHandling;
 use crate::core::{Feature, Features};
 use crate::util::GlobalContext;
-use crate::util::context::WarningHandling;
 
 #[derive(Clone, Debug)]
 pub struct Lint {
