+++ src/lib.rs
@@ -53,7 +53,7 @@
 //!   [rules][diagnostics::rules].
 //! - [`util`]:
 //!   This directory contains generally-useful utility modules.
-//! - [`util::context`]:
+//! - [`context`]:
 //!   This directory contains the global application context.
 //!   This includes the config parser which makes heavy use of
 //!   [serde](https://serde.rs/) to merge and translate config values.
@@ -127,7 +127,7 @@
 //!     - `src/*/*`: Extracted from `*.crate` by [`sources::registry::RegistrySource`]
 //!   - `git/`: Git source cache.  See [`sources::git`].
 //! - `**/.cargo/config.toml`: Environment dependent (env variables, files) configuration.  See
-//!   [`util::context`]
+//!   [`context`]
 //!
 //! ## Contribute to Cargo documentations
 //!
