+++ src/ops/cargo_compile/mod.rs
@@ -47,6 +47,7 @@ use crate::compiler::{CompileKind, CompileTarget, RustcTargetData, Unit};
 use crate::compiler::{CrateType, TargetInfo, apply_env_config, standard_lib};
 use crate::compiler::{DefaultExecutor, Executor, UnitInterner};
 use crate::compiler::{DepKindSet, UnitIndex};
+use crate::context::{GlobalContext, WarningHandling};
 use crate::core::profiles::Profiles;
 use crate::core::{PackageId, PackageSet, SourceId, TargetKind, Workspace};
 use crate::drop_println;
@@ -55,7 +56,6 @@ use crate::ops::resolve::{SpecsAndResolvedFeatures, WorkspaceResolve};
 use crate::resolver::features::{self, CliFeatures, FeaturesFor};
 use crate::resolver::{ForceAllTargets, HasDevUnits, Resolve};
 use crate::util::BuildLogger;
-use crate::util::context::{GlobalContext, WarningHandling};
 use crate::util::interning::InternedString;
 use crate::util::log_message::LogMessage;
 use crate::util::{CargoResult, StableHasher};
