+++ src/ops/cargo_update.rs
@@ -1,3 +1,4 @@
+use crate::context::GlobalContext;
 use crate::core::Registry as _;
 use crate::core::dependency::Dependency;
 use crate::core::registry::PackageRegistry;
@@ -10,7 +11,6 @@ use crate::resolver::features::{CliFeatures, HasDevUnits};
 use crate::sources::IndexSummary;
 use crate::sources::source::QueryKind;
 use crate::util::cache_lock::CacheLockMode;
-use crate::util::context::GlobalContext;
 use crate::util::toml_mut::dependency::{MaybeWorkspace, Source};
 use crate::util::toml_mut::manifest::LocalManifest;
 use crate::util::toml_mut::upgrade::upgrade_requirement;
