+++ src/ops/common_for_install_and_uninstall.rs
@@ -13,6 +13,7 @@ use ops::FilterRule;
 use serde::{Deserialize, Serialize};
 
 use crate::compiler::{DirtyReason, Freshness};
+use crate::context::{ConfigRelativePath, Definition};
 use crate::core::{Dependency, FeatureValue, Package, PackageId, SourceId};
 use crate::core::{PackageSet, Target};
 use crate::ops::{self, CompileFilter, CompileOptions};
@@ -21,7 +22,6 @@ use crate::sources::PathSource;
 use crate::sources::source::{QueryKind, Source, SourceMap};
 use crate::util::GlobalContext;
 use crate::util::cache_lock::CacheLockMode;
-use crate::util::context::{ConfigRelativePath, Definition};
 use crate::util::errors::CargoResult;
 use crate::util::{FileLock, Filesystem};
