+++ src/ops/registry/cargo_yank.rs
@@ -8,8 +8,8 @@ use anyhow::bail;
 use cargo_credential::Operation;
 use cargo_credential::Secret;
 
+use crate::context::GlobalContext;
 use crate::core::Workspace;
-use crate::util::context::GlobalContext;
 use crate::util::errors::CargoResult;
 use crate::util::important_paths::find_root_manifest_for_wd;
