+++ src/util/credential/paseto.rs
@@ -13,10 +13,10 @@ use time::{OffsetDateTime, format_description::well_known::Rfc3339};
 use url::Url;
 
 use crate::{
-    GlobalContext,
+    GlobalContext, context,
     core::SourceId,
     ops::RegistryCredentialConfig,
-    util::{auth::registry_credential_config_raw, command_prelude::opt, context},
+    util::{auth::registry_credential_config_raw, command_prelude::opt},
 };
 
 /// The main body of an asymmetric token as describe in RFC 3231.
