+++ src/util/credential/token.rs
@@ -5,10 +5,8 @@ use cargo_credential::{Action, CacheControl, Credential, CredentialResponse, Err
 use url::Url;
 
 use crate::{
-    GlobalContext,
-    core::SourceId,
-    ops::RegistryCredentialConfig,
-    util::{auth::registry_credential_config_raw, context},
+    GlobalContext, context, core::SourceId, ops::RegistryCredentialConfig,
+    util::auth::registry_credential_config_raw,
 };
 
 pub struct TokenCredential<'a> {
