+++ src/util/network/http.rs
@@ -15,8 +15,8 @@ use tracing::trace;
 
 use crate::CargoResult;
 use crate::GlobalContext;
-use crate::util::context::SslVersionConfig;
-use crate::util::context::SslVersionConfigRange;
+use crate::context::SslVersionConfig;
+use crate::context::SslVersionConfigRange;
 
 /// Creates a new HTTP handle with appropriate global configuration for cargo.
 pub fn http_handle(gctx: &GlobalContext) -> CargoResult<Easy> {
